Move all name parsing to the crawler
